Initial story idea
The first record I have of this story idea is a note I jotted to myself
in 2007. It's pretty clear that from the beginning I considered this a
joke story, a one-trick pony: the title itself is the story's main
punchline.

It wasn't until 2010 that I'd actually try to do something with the idea.