Sketching out more notes
Did some writing in the morning the next day.

Clearly I was still juggling a few different story ideas.

The terrible handwriting on the last 2 paragraphs indicate to me that
they were done on the bus on the way to campus.

It kinda freaks me out that the opening lines of the piece came out in a
rush, fully-formed, the voice already intact... and that they were composed
on a jittery rain-drenched bus while I was (probably) sandwiched between
commuters.

I'd been winding my way throughtdifferent aspects of the story in my
notebook that morning, and I suppose the pump had been primed... so at some
point during the commute, my subconscious took the fragments I'd been
feeding it and spit out something new.

I'm just glad I jotted it down.